Lazily renamed and moved base.Response and base.Result
Lazily moved base.Response to module results and renamed it to "ResultSet"
Lazily moved base.Result to module results, integrated base.parseAttrs
Introduced models.APIResponse (inheriting from requests.Response)
Integrated response creation into api.API
Introduced module exceptions and new InvalidIcinga2ApiResponseError
Introduced some (yet) empty classes
