+++ icinga2api_py/models.py
@@ -2,7 +2,7 @@
 """In this module are objects important for functionality."""
 
 import logging
-from requests import Request
+from requests import Request, Response
 
 
 class APIRequest(Request):
@@ -40,5 +40,25 @@ class APIRequest(Request):
 		# Take environment variables into account
 		settings = self.api.merge_environment_settings(request.url, None, None, None, None)
 		r = self.api.send(request, **settings)
-		return r if not self.api.response_parser else self.api.response_parser(r)
+		return self.api.create_response(r)
 
+
+class APIResponse(Response):
+	"""Represents a response from the Icinga2 API."""
+	def __init__(self):
+		super().__init__()
+
+	@staticmethod
+	def from_response(response):
+		"""Create a response from a response and return it."""
+		res = APIResponse()
+		res.__setstate__(response.__getstate__)
+		return res
+
+	def json(self, **kwargs):
+		"""JSON encoded content of the response (if any). Returns None on error."""
+		try:
+			return super().json(**kwargs)
+		except ValueError:
+			# No valid JSON encoding
+			return None
